+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Add `Azure::Core::Http::Request` constructor overload to support payload and non-buffered response.
+
 ### Breaking Changes
 
 ### Bugs Fixed